project_templates/ml/gpt-zap/requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.8.3
torchmetrics==0.11.0
tiktoken>=0.2.0
torchtext==0.14.1
xformers==0.0.16

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
rich            # beautiful text formatting in terminal
pytest          # tests